uiディレクトリを追加
